requirements.txt
Flask==2.1.2
Pillow==9.4.0
python_telegram_bot==13.14
requests==2.27.1
telegram==0.0.1
waitress==2.1.2
MarkupSafe==2.0.1
GoogleBard==1.0.0
